backend/requirements.txt
mysql-connector-python
flask
requests
pandas
httpx
transformers
torch
datasets
accelerate
loralib
scipy
bitsandbytes==0.41.1
pytz
kaleido
scikit-learn
plotly

git+https://github.com/huggingface/transformers.git@main
git+https://github.com/huggingface/peft.git@main
